Split StatusSchema and RangeLock from ManagementAPI
