+++ fdbclient/include/fdbclient/ManagementAPI.h
@@ -31,7 +31,6 @@ standard API and some knowledge of the contents of the system key space.
 #include <map>
 #include "fdbclient/GenericManagementAPI.h"
 #include "fdbclient/NativeAPI.actor.h"
-#include "fdbclient/RangeLock.h"
 #include "fdbclient/ReadYourWrites.h"
 #include "fdbclient/DatabaseConfiguration.h"
 #include "fdbclient/MonitorLeader.h"
@@ -357,40 +356,6 @@ Future<Optional<BulkDumpOwnerInfo>> getBulkLoadOwner(Database cx, UID jobId);
 
 // ==================== End Progress Tracking ====================
 
-// Persist a rangeLock owner to database metadata
-// A range can only be locked by a registered owner
-Future<Void> registerRangeLockOwner(Database cx, RangeLockOwnerName ownerUniqueID, std::string description);
-
-// Remove an owner form the database metadata
-Future<Void> removeRangeLockOwner(Database cx, RangeLockOwnerName ownerUniqueID);
-
-// Get all registered rangeLock owner
-AsyncResult<std::vector<RangeLockOwner>> getAllRangeLockOwners(Database cx);
-
-// Get a rangeLock owner by ownerUniqueID
-Future<Optional<RangeLockOwner>> getRangeLockOwner(Database cx, RangeLockOwnerName ownerUniqueID);
-
-// Block write traffic to a user range (the input range must be within normalKeys).
-// One transaction can call releaseExclusiveReadLockOnRange at most for one time.
-Future<Void> takeExclusiveReadLockOnRange(Transaction* tr, KeyRange range, RangeLockOwnerName ownerUniqueID);
-
-Future<Void> takeExclusiveReadLockOnRange(Database cx, KeyRange range, RangeLockOwnerName ownerUniqueID);
-
-// Unblock a user range (the input range must be within normalKeys).
-// One transaction can call releaseExclusiveReadLockOnRange at most for one time.
-Future<Void> releaseExclusiveReadLockOnRange(Transaction* tr, KeyRange range, RangeLockOwnerName ownerUniqueID);
-
-Future<Void> releaseExclusiveReadLockOnRange(Database cx, KeyRange range, RangeLockOwnerName ownerUniqueID);
-
-// Get locked ranges within the input range (the input range must be within normalKeys)
-Future<std::vector<std::pair<KeyRange, RangeLockState>>> findExclusiveReadLockOnRange(
-    Database cx,
-    KeyRange range,
-    Optional<RangeLockOwnerName> ownerName = Optional<RangeLockOwnerName>());
-
-// Clear all exclusive read lock by the input user. Not transactional.
-Future<Void> releaseExclusiveReadLockByUser(Database cx, RangeLockOwnerName ownerUniqueID);
-
 Future<Void> printHealthyZone(Database cx);
 Future<bool> clearHealthyZone(Database cx, bool printWarning = false, bool clearSSFailureZoneString = false);
 Future<bool> setHealthyZone(Database cx, StringRef zoneId, double seconds, bool printWarning = false);
@@ -399,15 +364,6 @@ Future<Void> waitForPrimaryDC(Database cx, StringRef dcId);
 
 // Gets the cluster connection string
 Future<Optional<ClusterConnectionString>> getConnectionString(Database cx);
-
-void schemaCoverage(std::string const& spath, bool covered = true);
-bool schemaMatch(json_spirit::mValue const& schema,
-                 json_spirit::mValue const& result,
-                 std::string& errorStr,
-                 Severity sev = SevError,
-                 bool checkCoverage = false,
-                 std::string path = std::string(),
-                 std::string schema_path = std::string());
 
 // execute payload in 'snapCmd' on all the coordinators, TLogs and
 // storage nodes
