+++ fdbclient/include/fdbclient/StatusSchema.h
@@ -1,0 +1,39 @@
+/*
+ * StatusSchema.h
+ *
+ * This source file is part of the FoundationDB open source project
+ *
+ * Copyright 2013-2026 Apple Inc. and the FoundationDB project authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#ifndef FDBCLIENT_STATUSSCHEMA_H
+#define FDBCLIENT_STATUSSCHEMA_H
+#pragma once
+
+#include <string>
+
+#include "fdbclient/json_spirit/json_spirit_value.h"
+#include "flow/Trace.h"
+
+void schemaCoverage(std::string const& spath, bool covered = true);
+bool schemaMatch(json_spirit::mValue const& schema,
+                 json_spirit::mValue const& result,
+                 std::string& errorStr,
+                 Severity sev = SevError,
+                 bool checkCoverage = false,
+                 std::string path = std::string(),
+                 std::string schema_path = std::string());
+
+#endif
