+++ fdbserver/workloads/BulkLoading.cpp
@@ -21,6 +21,7 @@
 #include "fdbclient/BulkLoading.h"
 #include "fdbclient/ManagementAPI.h"
 #include "fdbclient/NativeAPI.actor.h"
+#include "fdbclient/RangeLock.h"
 #include "fdbclient/SystemData.h"
 #include "fdbserver/core/BulkLoadUtil.h"
 #include "fdbserver/core/RocksDBCheckpointUtils.h"
