+++ fdbserver/workloads/ChangeConfig.cpp
@@ -21,6 +21,7 @@
 #include "fdbclient/ManagementAPI.h"
 #include "fdbclient/NativeAPI.actor.h"
 #include "fdbclient/Schemas.h"
+#include "fdbclient/StatusSchema.h"
 #include "fdbserver/tester/workloads.h"
 #include "fdbrpc/simulator.h"
