+++ fdbserver/workloads/StatusWorkload.cpp
@@ -24,7 +24,7 @@
 #include "fdbclient/StatusClient.h"
 #include "flow/UnitTest.h"
 #include "fdbclient/Schemas.h"
-#include "fdbclient/ManagementAPI.h"
+#include "fdbclient/StatusSchema.h"
 
 struct StatusWorkload : TestWorkload {
 	static constexpr auto NAME = "Status";
